requirements.txt
httpx>=0.27.0,<1
python-dotenv>=1.0.0,<2
pytest>=8.0.0,<9
